Remove TOID and Load Default Point Value
It was decided that it'd be easier to store the machine name of the
transaction operation the config is associated with as the ID as opposed
to storing it in a seperate value. When editing a transaction
operation, the form will now load with a default point value from the
config if there already exists a config object associated with the
operation.