User Points Payment Method Foundation
Got the payment method for user points in a testing state. A user can
select the payment method of user points and complete their purchase
using that payment method. Right now the payment method only prints out
test information, it must be changed to execute a transaction on a
user's point balance.