Currency Field Formatters and View Field
Added custom field formatters for displaying the TUP currency as the
number of points plus "points" appended on to it, rather than putting in
the currency symbol. To fix the total area in the cart view, I had to
add a new global view field that did the same thing but for the totals.
I had to do it this way because the commerce module hard codes the total
area to use the order_summary formatter, which displays currency codes.