Multiple Transaction Bug Fix
I fixed a bug causing multiple operations to not be executed when
selected from the monthly report form on a tenant. There is a new bug
where a transaction's points are not being applied if 3 or more
transactions are executed on the same user at the same time. That will
need to be fixed in a separate commit.